+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2021-03-15
+* Added `st2chatops` support and service startup script. (#206)
+
 ## 2021-03-13
 * Switch to using `latest` tag for st2 Docker images (#222)